example/requirements.txt
Django
djangorestframework
twine
wheel
drf-simple-invite
factory_boy
coverage
gunicorn
newrelic
django-storages
boto3
nose